Rename indexfile package to boltdb
