+++ pkg/storage/stores/shipper/boltdb/index.go
@@ -1,4 +1,4 @@
-package indexfile
+package boltdb
 
 import (
 	"context"
