+++ pkg/storage/stores/shipper/index/table.go
@@ -18,7 +18,7 @@ import (
 	"github.com/grafana/loki/pkg/storage/chunk/client/local"
 	chunk_util "github.com/grafana/loki/pkg/storage/chunk/client/util"
 	"github.com/grafana/loki/pkg/storage/stores/series/index"
-	indexfile "github.com/grafana/loki/pkg/storage/stores/shipper/boltdb"
+	"github.com/grafana/loki/pkg/storage/stores/shipper/boltdb"
 	shipper_util "github.com/grafana/loki/pkg/storage/stores/shipper/util"
 	util_log "github.com/grafana/loki/pkg/util/log"
 )
@@ -172,7 +172,7 @@ func (lt *Table) Snapshot() error {
 	return nil
 }
 
-func (lt *Table) ForEach(_ context.Context, callback func(boltdb *bbolt.DB) error) error {
+func (lt *Table) ForEach(_ context.Context, callback func(b *bbolt.DB) error) error {
 	lt.dbSnapshotsMtx.RLock()
 	defer lt.dbSnapshotsMtx.RUnlock()
 
@@ -324,7 +324,7 @@ func (lt *Table) handoverIndexesToShipper(force bool) ([]string, error) {
 			continue
 		}
 
-		err = lt.indexShipper.AddIndex(lt.name, "", indexfile.BoltDBToIndexFile(db, lt.buildFileName(name)))
+		err = lt.indexShipper.AddIndex(lt.name, "", boltdb.BoltDBToIndexFile(db, lt.buildFileName(name)))
 		if err != nil {
 			return nil, err
 		}
@@ -361,7 +361,7 @@ func loadBoltDBsFromDir(dir string, metrics *metrics) (map[string]*bbolt.DB, err
 		}
 		fullPath := filepath.Join(dir, entry.Name())
 
-		if strings.HasSuffix(entry.Name(), indexfile.TempFileSuffix) || strings.HasSuffix(entry.Name(), snapshotFileSuffix) {
+		if strings.HasSuffix(entry.Name(), boltdb.TempFileSuffix) || strings.HasSuffix(entry.Name(), snapshotFileSuffix) {
 			// If an ingester is killed abruptly in the middle of an upload operation it could leave out a temp file which holds the snapshot of db for uploading.
 			// Cleaning up those temp files to avoid problems.
 			if err := os.Remove(fullPath); err != nil {
